+++ classes/backup.php
@@ -21,6 +21,9 @@
  * @copyright 2016 Lafayette College ITS
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
+
+namespace local_course_template;
+
 defined('MOODLE_INTERNAL') || die();
 
 require_once($CFG->dirroot . '/lib/filestorage/mbz_packer.php');
@@ -34,7 +37,7 @@ require_once($CFG->dirroot . '/backup/util/includes/restore_includes.php');
  * @copyright 2016 Lafayette College ITS
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-class local_course_template_backup {
+class backup {
     /**
      * Creates a backup for the given course and extracts it to temporary file storage.
      *
@@ -49,18 +52,18 @@ class local_course_template_backup {
         if ($storedfile === false) {
 
             // Instantiate controller.
-            $bc = new backup_controller(
-                \backup::TYPE_1COURSE, $courseid, backup::FORMAT_MOODLE, backup::INTERACTIVE_NO, backup::MODE_GENERAL, 2);
+            $bc = new \backup_controller(
+                \backup::TYPE_1COURSE, $courseid, \backup::FORMAT_MOODLE, \backup::INTERACTIVE_NO, \backup::MODE_GENERAL, 2);
 
             // Run the backup.
-            $bc->set_status(backup::STATUS_AWAITING);
+            $bc->set_status(\backup::STATUS_AWAITING);
             $bc->execute_plan();
             $result = $bc->get_results();
             $bc->destroy();
 
             // Store the backup.
             $file = $result['backup_destination'];
-            $context = context_course::instance($courseid);
+            $context = \context_course::instance($courseid);
             $timestamp = time();
             $fs = get_file_storage();
             $filerecord = array(
@@ -79,7 +82,7 @@ class local_course_template_backup {
         }
 
         // Extract the backup.
-        $packer = new mbz_packer();
+        $packer = new \mbz_packer();
         $storedfile->extract_to_pathname($packer, "$CFG->tempdir/backup/$courseid");
         return $courseid;
     }
@@ -93,13 +96,13 @@ class local_course_template_backup {
      * @param int $courseid the courseid of the template course
      * @return \stored_file|boolean
      */
-    private static function get_cached_course($courseid) {
+    public static function get_cached_course($courseid) {
         $enablecaching = get_config('local_course_template', 'enablecaching');
         if (empty($enablecaching) || $enablecaching == 0) {
             return false;
         }
-        $context = context_course::instance($courseid);
-        $cache = cache::make('local_course_template', 'backups');
+        $context = \context_course::instance($courseid);
+        $cache = \cache::make('local_course_template', 'backups');
         $storedfile = $cache->get($context->id);
         return $storedfile;
     }
@@ -113,12 +116,12 @@ class local_course_template_backup {
      * @param int $contextid the contextid of the course
      * @param \storedfile the stored file of the backup
      */
-    private static function set_cached_course($contextid, $storedfile) {
+    public static function set_cached_course($contextid, $storedfile) {
         $enablecaching = get_config('local_course_template', 'enablecaching');
         if (empty($enablecaching) || $enablecaching == 0) {
             return;
         }
-        $cache = cache::make('local_course_template', 'backups');
+        $cache = \cache::make('local_course_template', 'backups');
         $cache->set($contextid, $storedfile);
     }
 
@@ -131,8 +134,8 @@ class local_course_template_backup {
      */
     public static function restore_backup($templateid, $courseid) {
         $admin = get_admin();
-        $rc = new restore_controller(
-            $templateid, $courseid, backup::INTERACTIVE_NO, backup::MODE_SAMESITE, $admin->id, backup::TARGET_EXISTING_ADDING);
+        $rc = new \restore_controller(
+            $templateid, $courseid, \backup::INTERACTIVE_NO, \backup::MODE_SAMESITE, $admin->id, \backup::TARGET_EXISTING_ADDING);
         self::apply_defaults($rc);
         if (!$rc->execute_precheck(true)) {
             return false;
@@ -168,7 +171,7 @@ class local_course_template_backup {
         foreach ($settings as $name => $value) {
             if ($rc->get_plan()->setting_exists($name)) {
                 $setting = $rc->get_plan()->get_setting($name);
-                if ($setting->get_status() == backup_setting::NOT_LOCKED) {
+                if ($setting->get_status() == \backup_setting::NOT_LOCKED) {
                     $setting->set_value($value);
                 }
             }