+++ classes/helper.php
@@ -21,6 +21,9 @@
  * @copyright 2016 Lafayette College ITS
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
+
+namespace local_course_template;
+
 defined('MOODLE_INTERNAL') || die();
 
 /**
@@ -30,7 +33,7 @@ defined('MOODLE_INTERNAL') || die();
  * @copyright 2016 Lafayette College ITS
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-class local_course_template_helper {
+class helper {
     /**
      * Applies the course template to the given course.
      *
@@ -38,20 +41,19 @@ class local_course_template_helper {
      * @return bool A status indicating success or failure
      */
     public static function template_course($courseid) {
-
         $templatecourseid = self::find_term_template($courseid);
         if ($templatecourseid == false) {
             return;
         }
 
         // Create and extract template backup file.
-        $backupid = \local_course_template_backup::create_backup($templatecourseid);
+        $backupid = backup::create_backup($templatecourseid);
         if (!$backupid) {
             return false;
         }
 
         // Restore the backup.
-        $status = \local_course_template_backup::restore_backup($backupid, $courseid);
+        $status = backup::restore_backup($backupid, $courseid);
         if (!$status) {
             return false;
         }
@@ -60,8 +62,8 @@ class local_course_template_helper {
         self::prune_news_forums($courseid);
 
         // Trigger custom event.
-        $systemcontext = context_system::instance();
-        $event = \local_course_template\event\template_copied::create([
+        $systemcontext = \context_system::instance();
+        $event = event\template_copied::create([
             'context' => $systemcontext,
             'other' => [
                 'courseid' => $courseid,
@@ -109,6 +111,7 @@ class local_course_template_helper {
                     }
                     return false;
                 } else {
+                    self::set_cached_course_id($shortname, $course->id);
                     return $course->id;
                 }
             } else {
@@ -129,12 +132,12 @@ class local_course_template_helper {
      * @param string $shortname the shortname of the template course
      * @return int|boolean
      */
-    private static function get_cached_course_id($shortname) {
+    public static function get_cached_course_id($shortname) {
         $enablecaching = get_config('local_course_template', 'enablecaching');
         if (empty($enablecaching) || $enablecaching == 0) {
             return false;
         }
-        $cache = cache::make('local_course_template', 'templates');
+        $cache = \cache::make('local_course_template', 'templates');
         $courseid = $cache->get($shortname);
         return $courseid;
     }
@@ -148,12 +151,12 @@ class local_course_template_helper {
      * @param string $shortname the course shortname
      * @param int $courseid the course id
      */
-    private static function set_cached_course_id($shortname, $courseid) {
+    public static function set_cached_course_id($shortname, $courseid) {
         $enablecaching = get_config('local_course_template', 'enablecaching');
         if (empty($enablecaching) || $enablecaching == 0) {
             return;
         }
-        $cache = cache::make('local_course_template', 'templates');
+        $cache = \cache::make('local_course_template', 'templates');
         $cache->set($shortname, $courseid);
     }