+++ classes/task/cleanup_task.php
@@ -50,22 +50,6 @@ class cleanup_task extends \core\task\scheduled_task {
         global $DB;
 
         // Find and prune template backups.
-        $cache = \cache::make('local_course_template', 'backups');
-        $backups = $DB->get_records('files', array('component' => 'local_course_template', 'filearea' => 'backup'));
-        $fs = get_file_storage();
-        foreach ($backups as $record) {
-            $file = $fs->get_file(
-                $record->contextid,
-                $record->component,
-                $record->filearea,
-                $record->itemid,
-                $record->filepath,
-                $record->filename
-            );
-            if ($file) {
-                $file->delete();
-                $cache->delete($record->contextid);
-            }
-        }
+        cache::clear();
     }
 }